requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
altgraph==0.17.2
async-generator==1.10
async-timeout==4.0.2
attrs==21.2.0
auto-py-to-exe==2.10.1
backports.zoneinfo==0.2.1
bidict==0.21.4
bottle==0.12.19
bottle-websocket==0.2.9
bscscan-python==2.0.0
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
colorlog==6.5.0
configparser==5.0.2
crayons==0.4.0
cryptography==35.0.0
dateparser==1.1.0
EasyProcess==0.3
Eel==0.12.4
et-xmlfile==1.1.0
eth-hash==0.3.2
eth-typing==2.2.2
excel2json-3==0.1.6
fake-useragent==0.1.11
Flask==2.0.2
Flask-Dropzone==1.6.0
Flask-FontAwesome==0.1.5
Flask-SocketIO==5.1.1
Flask-Uploads==0.2.1
Flask-WTF==0.15.1
frozenlist==1.2.0
future==0.18.2
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.2
h11==0.12.0
ibm-cloud-sdk-core==3.11.3
ibm-watson==5.3.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
MouseInfo==0.1.3
multidict==5.2.0
mysql==0.0.3
mysqlclient==2.1.0
ntplib==0.4.0
numpy==1.21.2
opencv-python==4.5.4.60
openpyxl==3.0.9
outcome==1.1.0
pandas==1.3.3
pefile==2021.9.3
Pillow==8.4.0
py3cw==0.0.32
pyaes==1.6.1
pyasn1==0.4.8
PyAutoGUI==0.9.53
pycparser==2.20
pyfiglet==0.8.post1
PyGetWindow==0.0.9
pygtail==0.11.1
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
PyJWT==2.2.0
PyMsgBox==1.0.9
pyOpenSSL==21.0.0
pyotp==2.6.0
pyperclip==1.8.2
PyRect==0.1.4
PyScreeze==0.1.28
pytesseract==0.3.8
python-binance==1.0.15
python-dateutil==2.8.2
python-dotenv==0.19.1
python-engineio==4.3.0
python-socketio==5.4.1
pytweening==1.0.4
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyVirtualDisplay==2.2
pywin32-ctypes==0.2.0
random-user-agent==1.0.1
regex==2021.11.10
requests==2.26.0
rsa==4.8
selenium==4.0.0b4
selenium-stealth==1.0.6
six==1.16.0
sniffio==1.2.0
solver==0.0.4
sortedcontainers==2.4.0
Telethon==1.24.0
toolz==0.11.2
trio==0.19.0
trio-websocket==0.9.2
tzdata==2021.5
tzlocal==4.1
ua-parser==0.10.0
ujson==5.1.0
undetected-chromedriver==3.0.3
urllib3==1.26.7
user-agents==2.2.0
webdriver-manager==3.5.0
websocket-client==1.1.0
websockets==9.1
Werkzeug==2.0.2
wget==3.2
whichcraft==0.6.1
wsproto==1.0.0
WTForms==2.3.3
xlrd==2.0.1
yarl==1.7.2
zope.event==4.5.0
zope.interface==5.4.0
